# perf(dracut-install): preload kmod resources for quicker module lookup